Rewrite and refactor. cc:@LTe thanks for all the fish ^_^ :beer:
Change formatter to a Class. This way multiple formatters can have different
view roots, and they'll keep that setting in an instance variable.

Reimplement RablRails' Library - just an hack for now, that'll be merged
upstream later. This version is disentangled by any Rails internals, and
uses Thread.current for the view root - so it's thread safe while keeping
the Singleton cache for compiled partials.

RablRails' Library should then be augmented with an adapter for Rails,
that extracts the view root from the app configuration, and not rely on
Rails' internals - so it can be easily used from outside Rails and all
the jazz currenly in the standalone renderer would not be required.

cc:@ccocchi what do you think?